[pipe-cli] Enable login without explicitly specified tokens - review fixes
